added stack navigation and redo home
