+++ front/App.js
@@ -3,28 +3,23 @@ import { Provider } from 'react-redux';
 
 import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import store from "./redux/store";
 import Home from './components/Home/Home';
-import Currencies from './components/Currencies/Currencies';
-import Account from './components/Account/Account';
 import Login from './components/Login/Login';
 import Register from "./components/Register/Register"
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-
-const Tab = createBottomTabNavigator();
+const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
     <Provider store={store}>
       <NavigationContainer>
-        <Tab.Navigator>
-          <Tab.Screen name="Login" component={Login}/>
-          <Tab.Screen name="Register" component={Register}/>
-          <Tab.Screen name="Home" component={Home}/>
-          <Tab.Screen name="Currencies" component={Currencies}/>
-          <Tab.Screen name="Account" component={Account}/>
-        </Tab.Navigator>
+        <Stack.Navigator initialRouteName='Login'>
+          <Stack.Screen name="Login" component={Login}/>
+          <Stack.Screen name="Register" component={Register}/>
+          <Stack.Screen name="Home" component={Home}/>
+        </Stack.Navigator>
       </NavigationContainer>
     </Provider>
   );
